Contenido texto de las secciones ya funciona
Se realizaron los cambios correspondientes al uso de json en el editor
de seccion, resta hacer funcionar la parte de archivos adjuntos.

+++ index.php
@@ -628,15 +628,14 @@ $app->post('/editar-seccion-post/:id/', function($id) use ($app) {//XXX Actualiz
     $validated = $validator->validate($_POST, $rules);
     if ($validated === true) {
         $seccion = Seccion::find($id);
-        $contenido = $seccion->contenido;
-        $contenido = (array) json_decode($contenido);
+        $contenido = (array) json_decode($seccion->contenido);
         if(!array_key_exists('data', $contenido)){
             $contenido = array(
                 'data' => $_POST['contenido'],
                 'files' => array()
             );
         } else {
-            
+            $contenido['data'] = $_POST['contenido'];
         }
         $seccion->contenido = json_encode($contenido);// $_POST['contenido'];
         $seccion->contenedor = $_POST['contenedor'];
@@ -730,7 +729,7 @@ $app->get('/relacion-aceptados/', function() use ($app) {
     //TODO Relacion Aceptados
     $app->render('relacion-aceptados.html');
 })->name('relacion-aceptados');
-//TODO por ver
+
 $app->get('/publicaciones/', function() use ($app) {
     $user['usuarios'] = Usuario::find_by_id('1');
     $data['publicaciones']= Publicacion::find_all_by_usuario_id($user['usuarios']->id);
@@ -744,10 +743,6 @@ $app->get('/egresados/', function() use ($app) {
 $app->get('/estadisticas/matriculacion/', function() use ($app) {
     $app->render('EstadisticaMatriculacion.html');
 })->name('matriculacion');
-
-$app->get('/registro-aspirante/', function() use ($app) {
-    //TODO Registro Aspirante
-})->name('registro-aspirante');
 
 /* =======================
  * ======= DOCENTE =======
@@ -1002,7 +997,7 @@ $app->get('/docente/publicaciones/', function() use ($app) {
           );
           $app -> flash("flash", $flash);
           $app->flashKeep();
-          $app->redirect($app->urlFor('registro-aspirante'));
+          $app->redirect($app->urlFor('registro-aspirantes'));
 
       } else {
            $flash = array(
@@ -1013,7 +1008,7 @@ $app->get('/docente/publicaciones/', function() use ($app) {
       );
       $app -> flash("flash", $flash);
      $app->flashKeep();
-     $app->redirect($app->urlFor('registro-aspirante'));
+     $app->redirect($app->urlFor('registro-aspirantes'));
      }
     }
  })->name('registro-aspirante-post');
@@ -3059,9 +3054,9 @@ $app->post('/actualiza-carrera/:id/', function($id) use ($app) {
 
 //
 $app->get('/borrar-carrera/:id/', function($id) use ($app) {
-//	$relaciones = UsuariosHerramientas::find_all_by_herramienta_id($id);
-//	$cant_relaciones = count($relaciones);
-//	if ($cant_relaciones == 0) {
+	$relaciones = Academico::find_all_by_carrera($id); 
+	$cant_relaciones = count($relaciones);
+	if ($cant_relaciones == 0){		
         $carrera = Carrera::find($id);
         $carrera->delete();
         $flash = array(
@@ -3073,18 +3068,18 @@ $app->get('/borrar-carrera/:id/', function($id) use ($app) {
         $app -> flash("flash", $flash);
         $app->flashKeep();
         $app->redirect($app->urlFor('CatCarrera'));
-//	}
-//	else {
-//		$flash = array(
-//			"title" => "OK",
-//			"msg" => "La Carrera esta relacionado, no se permite la eliminación.",
-//			"type" => "info",
-//			"fade" => 1
-//		);
-//		$app -> flash("flash", $flash);
-//		$app->flashKeep();
-//		$app->redirect($app->urlFor('CatHerramienta'));
-//	}
+	}
+	else {
+		$flash = array(
+			"title" => "OK",
+			"msg" => "La Carrera esta relacionada, no se permite la eliminación.",
+			"type" => "info",
+			"fade" => 1
+		);
+		$app -> flash("flash", $flash);
+		$app->flashKeep();
+		$app->redirect($app->urlFor('CatCarrera'));
+	}
 })->name('borrar-carrera');
 
 //*******
@@ -3107,7 +3102,7 @@ $app->post('/nueva-titulacion/', function() use ($app) {
     $post = $_POST = $validator->filter($_POST, $filters);
     $validated = $validator->validate($_POST, $rules);
     if ($validated === TRUE) {
-        $titulacion = new Formas_titulacion();
+		$titulacion = new FormaTitulacion();
         $titulacion->nombre = $_POST['nombre'];
         $titulacion->save();
         $flash = array(
@@ -3145,7 +3140,7 @@ $app->post('/actualiza-titulacion/:id/', function($id) use ($app) {
     $post = $_POST = $validator->filter($_POST, $filters);
     $validated = $validator->validate($_POST, $rules);
     if ($validated === TRUE) {
-        $titulacion = Formas_titulacion::find($id);
+		$titulacion = FormaTitulacion::find($id);
         $titulacion -> nombre = $_POST['nombre-edit'];
         $titulacion -> save();
         $flash = array(
@@ -3173,11 +3168,11 @@ $app->post('/actualiza-titulacion/:id/', function($id) use ($app) {
 
 //
 $app->get('/borrar-titulacion/:id/', function($id) use ($app) {
-//	$relaciones = UsuariosHerramientas::find_all_by_herramienta_id($id);
-//	$cant_relaciones = count($relaciones);
-//	if ($cant_relaciones == 0) {
-        $carrera = Formas_titulacion::find($id);
-        $carrera->delete();
+	$relaciones = Academico::find_all_by_forma_titulacion($id); 
+	$cant_relaciones = count($relaciones);
+	if ($cant_relaciones == 0){		
+		$titulacion = FormaTitulacion::find($id);
+		$titulacion->delete();
         $flash = array(
             "title" => "OK",
             "msg" => "La Forma de Titulación ha sido borrada correctamente.",
@@ -3186,19 +3181,19 @@ $app->get('/borrar-titulacion/:id/', function($id) use ($app) {
         );
         $app -> flash("flash", $flash);
         $app->flashKeep();
-        $app->redirect($app->urlFor('CatCarrera'));
-//	}
-//	else {
-//		$flash = array(
-//			"title" => "OK",
-//			"msg" => "La Carrera esta relacionado, no se permite la eliminación.",
-//			"type" => "info",
-//			"fade" => 1
-//		);
-//		$app -> flash("flash", $flash);
-//		$app->flashKeep();
-//		$app->redirect($app->urlFor('CatHerramienta'));
-//	}
+		$app->redirect($app->urlFor('CatFormaTitulacion'));
+	}
+	else {
+		$flash = array(
+			"title" => "OK",
+			"msg" => "La Forma de Tirulación esta relacionado, no se permite la eliminación.",
+			"type" => "info",
+			"fade" => 1
+		);
+		$app -> flash("flash", $flash);
+		$app->flashKeep();
+		$app->redirect($app->urlFor('CatFormaTitulacion'));
+	}
 })->name('borrar-titulacion');
 
 /* =======================
@@ -3209,8 +3204,20 @@ $app->get('/(:slug/)', function ($slug = "") use ($app) {
     if ($slug != "") {
         $seccion = Seccion::find_by_slug($slug);
         if ($seccion) {
-            $seccion->contenido = replace_hashes($seccion->contenido);
-            $data['seccion'] = $seccion;
+            $contenido = (array) json_decode($seccion->contenido);
+            $proseccion = array(
+                'id' => $seccion->id,
+                'nombre' => $seccion->nombre,
+                'slug' => $seccion->slug,
+                'contenedor' => $seccion->contenedor,
+                'orden' => $seccion->orden,
+                'contenido' => replace_hashes($contenido['data']),
+                'files' => $contenido['files'],
+                'actualizado' => $seccion->actualizado
+            );
+            $data['seccion'] = $proseccion;
+#            $seccion->contenido = replace_hashes($seccion->contenido);
+#            $data['seccion'] = $seccion;
         }
     }
     $app->render('index.html', $data);
